\name{checkNonEmpty}
\alias{checkNonEmpty}
\title{Checks if vector or data set contains at least one element or column.
Note that for data frame it will NOT check if it has 1 or more rows.}
\usage{
  checkNonEmpty(data)
}
\arguments{
  \item{data}{vecotr to test}
}
\description{
  Checks if vector or data set contains at least one
  element or column. Note that for data frame it will NOT
  check if it has 1 or more rows.
}

