% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeKmeans.R
\name{computeClusterSample}
\alias{computeClusterSample}
\title{Random sample of clustered data}
\usage{
computeClusterSample(channel, km, sampleFraction, sampleSize, scaled = FALSE,
  includeId = FALSE, test = FALSE)
}
\arguments{
\item{channel}{connection object as returned by \code{\link{odbcConnect}}.}

\item{km}{result of k-means clustering obtained with \code{computeKmeans}.}

\item{sampleFraction}{one or more sample fractions to use in the sampling of data. (multipe 
sampling fractions are not yet supported.)}

\item{sampleSize}{total sample size (applies only when \code{sampleFraction} is missing).}

\item{scaled}{logical indicates if function uses original (default) or scaled variables.}

\item{includeId}{logical indicates if sample should include the key uniquely identifying
each data row.}

\item{test}{logical: if TRUE show what would be done, only (similar to parameter \code{test} in \pkg{RODBC} 
functions: \link{sqlQuery} and \link{sqlSave}).}
}
\description{
Random sample of clustered data
}
\examples{

computeClusterSample(NULL, km, 0.01, test=TRUE)
}

