% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGraph.R
\name{computeGraph}
\alias{computeGraph}
\title{Materialize Aster graph as network object in R.}
\usage{
computeGraph(channel, graph, v = NULL, vertexWhere = graph$vertexWhere,
  edgeWhere = graph$edgeWhere, allTables = NULL, test = FALSE)
}
\arguments{
\item{channel}{connection object as returned by \code{\link{odbcConnect}}}

\item{graph}{an object of class \code{'toagraph'} referencing graph 
tables in Aster database.}

\item{v}{vector or list of keys (corresponds to \code{vertex.names} attribute) of the vertices 
to include in the graph. When not null this guarentees that no vertices outside of this list
and no edges between vertices outsidef of this list will be included in the resulting network.}

\item{vertexWhere}{optionally, a \code{SQL WHERE} clause to subset vertex tablet. When not \code{NULL}
it overrides \code{vertexWhere} condition from the \code{graph}.}

\item{edgeWhere}{optionally, a \code{SQL WHERE} clause to subset edge table. When not \code{NULL}
it overrides \code{edgeWhere} condition from the \code{graph}.}

\item{allTables}{pre-built information about existing tables.}

\item{test}{logical: if TRUE show what would be done, only (similar to parameter \code{test} in \pkg{RODBC} 
functions: \link{sqlQuery} and \link{sqlSave}).}
}
\description{
Results in \code{\link{network}} object representation of the graph 
stored in Aster tables. Usually in Aster database a graph is represented
using a pair of vertice and edge tables.
}
\details{
Use caution when computing network objects stored in Aster with this function 
as data may include considerable amount of vertices and edges which are too large to
load into a memory.
}

