% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeKmeans.R
\name{computeKmeans}
\alias{computeKmeans}
\title{Perform k-means clustering on the table.}
\usage{
computeKmeans(channel, tableName, tableInfo, id, include = NULL,
  except = NULL, centers = 6, aggregates = "COUNT(*) cnt", scale = TRUE,
  idAlias = "id", where = NULL, scaledTableName = NULL,
  centroidTableName = NULL, schema = NULL, test = FALSE)
}
\arguments{
\item{channel}{connection object as returned by \code{\link{odbcConnect}}.}

\item{tableName}{Aster table name.}

\item{tableInfo}{pre-built summary of data to use (require when \code{test=TRUE}). See \code{\link{getTableSummary}}.}

\item{id}{column name or SQL expression containing unique table key.}

\item{include}{a vector of column names with variables (must be numeric). Model never contains variables other than in the list.}

\item{except}{a vector of column names to exclude from variables. Model never contains variables from the list.}

\item{centers}{either the number of clusters, say k, or a set of initial (distinct) cluster centres (not supported yet). 
If a number, a random set of (distinct) rows in x is chosen as the initial centres.}

\item{aggregates}{vector with SQL aggregates to compute for each cluster identified by this function. Aggregate may have 
optional aliases like in \code{"AVG(era) avg_era"}. Subsequently, use in \code{createClusterPlot} as cluster properties.}

\item{scale}{logical if TRUE then scale each variable in-database before clustering. Currently scaling to 0 mean and unit
standard deviation only is supported.}

\item{idAlias}{SQL alias for table id. This is required when SQL expression is given for \code{id}.}

\item{where}{specifies criteria to satisfy by the table rows before applying
computation. The creteria are expressed in the form of SQL predicates (inside
\code{WHERE} clause).}

\item{scaledTableName}{name of Aster table with results of scaling}

\item{centroidTableName}{name of Aster table with centroids found by kmeans}

\item{schema}{name of Aster schema tables \code{scaledTableName} and \code{centroidTableName} belong.}

\item{test}{logical: if TRUE show what would be done, only (similar to parameter \code{test} in \pkg{RODBC} 
functions: \link{sqlQuery} and \link{sqlSave}).}
}
\description{
Perform k-means clustering on the table.
}
\examples{
computeKmeans(NULL, "batting", battingInfo, 
  id="playerid || '-' || stint || '-' || teamid || '-' || yearid", include=c('g','g_batting','ab','r','h','hr','rbi'),
  scaledTableName='kmeans_test_scaled', centroidTableName='kmeans_test_centroids', schema='baseball',
  test=TRUE)
  
if(interactive()) {
  computeKmeans(NULL, "batting", battingInfo, 
  id="playerid || '-' || stint || '-' || teamid || '-' || yearid", include=c('g','g_batting','ab','r','h','hr','rbi'))
}
}

