% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingKmeans.R
\name{createCentroidPlot}
\alias{createCentroidPlot}
\title{Create centroid plot}
\usage{
createCentroidPlot(centroids, format = c("line"), clusterId = "clusterid",
  baseSize = 12, baseFamily = "serif", title = paste("Cluster Centroids",
  format, "Plot"), xlab = ifelse(format == "line", "cluster", "variable"),
  ylab = ifelse(format == "heatmap", "cluster", "scaled value"),
  legendPosition = ifelse(format == "bar", "none", "right"),
  defaultTheme = ggthemes::theme_tufte(base_size = baseSize, base_family =
  baseFamily, ticks = FALSE), themeExtra = NULL)
}
\arguments{
\item{centroids}{Kmeans centroids returned by \code{computeKmeans}.}

\item{format}{type of plot to use: line, bar, bar_dodge, bar_facet (same as bar) or heatmap.}

\item{clusterId}{name of a column with cluster id.}

\item{baseSize}{\code{\link{theme}} base font size.}

\item{baseFamily}{\code{\link{theme}} base font family.}

\item{title}{plot title.}

\item{xlab}{a label for the x axis, defaults to a description of x.}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{legendPosition}{the position of legends. ("left", "right", "bottom", "top", or two-element numeric 
vector). "none" is no legend.}

\item{defaultTheme}{plot theme to use: \code{\link[ggthemes]{theme_tufte}} is default.}

\item{themeExtra}{any additional \code{\link[ggplot2]{theme}} settings that override default theme.}
}
\value{
ggplot object
}
\description{
Create plot of cluster centroids produced by clustering function like k-means.
Plot types available are line plot, bar plot, or heatmap. Use parameter \code{format}
to specify which plot to use.
}

