% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingKmeans.R
\name{createClusterPairsPlot}
\alias{createClusterPairsPlot}
\title{Create cluster variable plot}
\usage{
createClusterPairsPlot(kms, clusterId = "clusterid",
  title = "Cluster Variable Pairs", baseSize = 12, baseFamily = "serif",
  ticks = TRUE, defaultTheme = ggthemes::theme_tufte(base_size = baseSize,
  base_family = baseFamily, ticks = ticks), themeExtra = theme(), ...)
}
\arguments{
\item{kms}{sample of points assigned to clusters}

\item{clusterId}{name of a column with cluster id.}

\item{title}{plot title.}

\item{baseSize}{\code{\link{theme}} base font size.}

\item{baseFamily}{\code{\link{theme}} base font family.}

\item{ticks}{\code{logical} Show axis ticks?}

\item{defaultTheme}{plot theme to use: \code{\link[ggthemes]{theme_tufte}} is default.}

\item{themeExtra}{any additional \code{\link[ggplot2]{theme}} settings that override default theme.}

\item{...}{other parameters being suplied to geom's \code{aes}}
}
\value{
ggplot object
}
\description{
Create cluster variable plot
}

