% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingKmeans.R
\name{createClusterPlot}
\alias{createClusterPlot}
\title{Create clusters' properties plot}
\usage{
createClusterPlot(aggregates, clusterId = "clusterid", baseSize = 12,
  baseFamily = "serif", title = paste("Cluster Properties Plot"),
  xlab = "cluster", ylab = "value", border = TRUE,
  defaultTheme = ggthemes::theme_tufte(base_size = baseSize, base_family =
  baseFamily, ticks = FALSE), themeExtra = NULL)
}
\arguments{
\item{aggregates}{lis of cluster properties}

\item{clusterId}{name of a column with cluster id.}

\item{baseSize}{\code{\link{theme}} base font size.}

\item{baseFamily}{\code{\link{theme}} base font family.}

\item{title}{plot title.}

\item{xlab}{a label for the x axis, defaults to a description of x.}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{border}{boolean indicates to use border around plotting area. In case of facets border is around each facet.}

\item{defaultTheme}{plot theme to use: \code{\link[ggthemes]{theme_tufte}} is default.}

\item{themeExtra}{any additional \code{\link[ggplot2]{theme}} settings that override default theme.}
}
\value{
ggplot object
}
\description{
Create clusters' properties plot
}

