% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getArbitraryPrecisionTypes}
\alias{getArbitraryPrecisionTypes}
\title{List Aster arbitrary precision number data types.}
\usage{
getArbitraryPrecisionTypes()
}
\value{
character vector with names of Aster arbitrary precision numeric data types
}
\description{
List Aster arbitrary precision number data types.
}
\examples{
getArbitraryPrecisionTypes()

}
\seealso{
\code{\link{getNumericTypes}}, \code{\link{getFloatingPointTypes}}, \code{\link{getIntegerTypes}}
}

