% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getCharacterTypes}
\alias{getCharacterTypes}
\title{List Aster character data types.}
\usage{
getCharacterTypes()
}
\value{
character vector with names of Aster character data types
}
\description{
List Aster character data types.
}
\examples{
getCharacterTypes()


}
\seealso{
\code{\link{getNumericTypes}}, \code{\link{getTemporalTypes}}, \code{\link{getTableSummary}}
}
