% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{getDiscretePaletteFactory}
\alias{getDiscretePaletteFactory}
\title{Generate discrete palette maker}
\usage{
getDiscretePaletteFactory(paletteName = "Set1")
}
\arguments{
\item{paletteName}{name of palette from \code{brewer.pal.info} in \code{RColorBrewer} package}
}
\value{
function (factory) that creates discrete palette with for number of colors
}
\description{
Generate discrete palette maker
}
\examples{
paletteMaker = getDiscretePaletteFactory("PuOr")
myPalette = paletteMaker(25)
}
\seealso{
\code{\link{getGradientPaletteFactory}}, \code{\link{colorRampPalette}}
}
