% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getFloatingPointTypes}
\alias{getFloatingPointTypes}
\title{List Aster floating piont numeric data types.}
\usage{
getFloatingPointTypes()
}
\value{
character vector with names of Aster floating point numeric data types
}
\description{
List Aster floating piont numeric data types.
}
\examples{
getFloatingPointTypes()

}
\seealso{
\code{\link{getNumericTypes}}, \code{\link{getIntegerTypes}}, \code{\link{getArbitraryPrecisionTypes}}
}

