% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{getGradientPaletteFactory}
\alias{getGradientPaletteFactory}
\title{Generate gradient palette maker}
\usage{
getGradientPaletteFactory(colors = c("black", "white"))
}
\arguments{
\item{colors}{pair of colors for gradient range (min, max): default is \code{c('black','white')}}
}
\value{
function (factory) that creates linear gradient palette for given number of colors
}
\description{
inspired by
http://stackoverflow.com/questions/13353213/gradient-of-n-colors-ranging-from-color-1-and-color-2
}
\examples{
paletteMaker = getGradientPaletteFactory(c("yellow","red"))
myPalette = paletteMaker(10)
}
\seealso{
\code{\link{getDiscretePaletteFactory}}, \code{\link{colorRampPalette}}
}

