% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getIntegerTypes}
\alias{getIntegerTypes}
\title{List Aster integer data types.}
\usage{
getIntegerTypes()
}
\value{
character vector with names of Aster integer data types
}
\description{
List Aster integer data types.
}
\examples{
getIntegerTypes()

}
\seealso{
\code{\link{getNumericTypes}}, \code{\link{getFloatingPointTypes}}, \code{\link{getArbitraryPrecisionTypes}}
}
