% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{getNumericTypes}
\alias{getNumericTypes}
\title{List Aster numeric data types.}
\usage{
getNumericTypes()
}
\value{
character vector with names of Aster numeric data types
}
\description{
List Aster numeric data types.
}
\examples{
getNumericTypes()
}
\seealso{
\code{\link{getCharacterTypes}}, \code{\link{getTemporalTypes}}, \code{\link{getTableSummary}}
}

