% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getTemporalTypes}
\alias{getTemporalTypes}
\title{List Aster temporal data types.}
\usage{
getTemporalTypes()
}
\value{
character vector with names of Aster temporal data types
}
\description{
List Aster temporal data types.
}
\examples{
getTemporalTypes()

}
\seealso{
\code{\link{getCharacterTypes}}, \code{\link{getCharacterTypes}}, \code{\link{getTableSummary}}
}

