% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getWindowFunction}
\alias{getWindowFunction}
\title{Determine window function to use}
\usage{
getWindowFunction(rankFunction)
}
\arguments{
\item{rankFunction}{one of rank function codes to map to one of SQL window
functions for ranking.}
}
\description{
Determine window function to use
}
