% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{makeFromClause}
\alias{makeFromClause}
\title{Make SQL FROM clause}
\usage{
makeFromClause(name, flag, alias = "t")
}
\arguments{
\item{name}{table or view name or a SQL query.}

\item{flag}{logical indicates if a table or a query is visible.
Special value \code{NA} indicates that \code{name} is
a SQL query.}

\item{alias}{query alias to use. Ignored if \code{name} is 
a table or a view.}
}
\description{
Make SQL FROM clause
}

