\name{makeOrderByClause}
\alias{makeOrderByClause}
\title{Helper function to construct SQL \code{ORDER BY} clause}
\usage{
  makeOrderByClause(order)
}
\arguments{
  \item{order}{vector with column names and optional order
  directives (\code{asc/desc}).}
}
\description{
  Helper function to construct SQL \code{ORDER BY} clause
}

