% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_empty}
\alias{theme_empty}
\title{Creates empty theme.}
\usage{
theme_empty(baseSize = 12, baseFamily = "")
}
\arguments{
\item{baseSize}{base font size}

\item{baseFamily}{base font family}
}
\description{
Good to use with slopegraphs.
}
\seealso{
\code{\link{createHistogram}} and other visualization functions that start with create.
}

