% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingKmeans.R
\name{createCentroidPlot}
\alias{createCentroidPlot}
\title{Create plot of cluster centroids.}
\usage{
createCentroidPlot(km, format = "line", groupByCluster = TRUE,
  clusters = rownames(km$centers), dims = colnames(km$centers),
  baseSize = 12, baseFamily = "serif", title = paste("Cluster Centroids",
  format, "Plot"), xlab, ylab = ifelse(format == "heatmap", "cluster",
  ifelse(!is.null(km$scale) && km$scale, "scaled value", "value")),
  legendPosition = ifelse(format == "bar", "none", "right"),
  coordFlip = FALSE, ticks = FALSE, defaultTheme = theme_tufte(base_size =
  baseSize, base_family = baseFamily, ticks = ticks), themeExtra = NULL)
}
\arguments{
\item{km}{an object of class \code{"toakmeans"} returned by \code{\link{computeKmeans}} or of class 
\code{"kmeans"} returned by \code{\link{kmeans}}.}

\item{format}{type of plot to use: \code{"line"}, \code{"bar"}, \code{"bar_dodge"} or \code{"heatmap"}.}

\item{groupByCluster}{logical: indicates if centroids are grouped by clusters or variables. \code{groupByCluster} 
has no effect when \code{format="heatmap"}.}

\item{clusters}{optional character vector with clusters to include. If it has named values then names are used for 
cluster labels. By default, all clusters are included. Numerical vector will be casted to character one as is.}

\item{dims}{optional vector with dimensions to include. By default, all dimensions are included.}

\item{baseSize}{\code{\link{theme}} base font size.}

\item{baseFamily}{\code{\link{theme}} base font family.}

\item{title}{plot title.}

\item{xlab}{a label for the x axis, defaults to a description of x.}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{legendPosition}{the position of legends. ("left", "right", "bottom", "top", or two-element numeric 
vector). "none" is no legend.}

\item{coordFlip}{logical flipped cartesian coordinates so that horizontal becomes vertical, and vertical horizontal (see 
\link{coord_flip}).}

\item{ticks}{\code{logical} Show axis ticks using default theme settings (see \code{defaultTheme})?}

\item{defaultTheme}{plot theme settings with default value \code{\link[ggthemes]{theme_tufte}}. More themes
are available here: \code{\link[ggplot2]{ggtheme}} (by \href{http://ggplot2.org/}{ggplot2}) 
and \code{\link[ggthemes]{ggthemes}}.}

\item{themeExtra}{any additional \code{\link[ggplot2]{theme}} settings that override default theme.}
}
\value{
ggplot object
}
\description{
Visualize centroids produced by clustering function like k-means.
Plots available are line plot, bar plot, or heatmap. Parameter \code{format}
specifies which one to create.
}
\examples{
if(interactive()){
# initialize connection to Lahman baseball database in Aster 
conn = odbcDriverConnect(connection="driver={Aster ODBC Driver};
                         server=<dbhost>;port=2406;database=<dbname>;uid=<user>;pwd=<pw>")
                         
km = computeKmeans(conn, "batting", centers=5, iterMax = 25,
                   aggregates = c("COUNT(*) cnt", "AVG(g) avg_g", "AVG(r) avg_r", "AVG(h) avg_h"),
                   id="playerid || '-' || stint || '-' || teamid || '-' || yearid", 
                   include=c('g','r','h'), scaledTableName='kmeans_test_scaled', 
                   centroidTableName='kmeans_test_centroids',
                   where="yearid > 2000")
createCentroidPlot(km)
# As clusters fluctuate from run to run cluster names may be inconsistent
createCentroidPlot(km, format="bar_dodge", 
                   clusters=c(Stars=3, Reserve=4), dims=c('g','h'))
createCentroidPlot(km, format="heatmap", coordFlip=TRUE)
}
}

