% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingKmeans.R
\name{createSilhouetteProfile}
\alias{createSilhouetteProfile}
\title{Create cluster silhouette profile plot.}
\usage{
createSilhouetteProfile(km, baseSize = 12, baseFamily = "serif",
  title = "Cluster Silhouette Profile (Histogram)",
  xlab = "Silhouette Value", ylab = "Count", ticks = FALSE,
  coordFlip = TRUE, defaultTheme = ggthemes::theme_tufte(base_size =
  baseSize, base_family = baseFamily, ticks = ticks), themeExtra = NULL)
}
\arguments{
\item{km}{an object of class \code{"toakmeans"} returned by \code{\link{computeKmeans}}.}

\item{baseSize}{\code{\link{theme}} base font size.}

\item{baseFamily}{\code{\link{theme}} base font family.}

\item{title}{plot title.}

\item{ticks}{\code{logical} Show axis ticks?}

\item{coordFlip}{logical flipped cartesian coordinates so that horizontal becomes vertical, and vertical horizontal (see 
\link{coord_flip}).}

\item{defaultTheme}{plot theme to use: \code{\link[ggthemes]{theme_tufte}} is default.}

\item{themeExtra}{any additional \code{\link[ggplot2]{theme}} settings that override default theme.}
}
\value{
ggplot object
}
\description{
Create cluster silhouette profile plot.
}

