% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showGraph.R
\name{showGraph}
\alias{showGraph}
\title{Plot an Aster graph object comprised of the vertice and edge tables.}
\usage{
showGraph(channel, graph, allTables = NULL, test = FALSE)
}
\arguments{
\item{channel}{connection object as returned by \code{\link{odbcConnect}}}

\item{graph}{an object of class \code{'toagraph'} referencing graph 
tables in Aster database.}

\item{allTables}{pre-built information about existing tables.}

\item{test}{logical: if TRUE show what would be done, only (similar to parameter \code{test} in \pkg{RODBC} 
functions: \link{sqlQuery} and \link{sqlSave}).}
}
\value{
a ggplot object
}
\description{
Function for obtaining and plotting graph (network) objects using ggplot2 and ggnet2 functions.
}
\examples{
if(interactive()) {
# initialize connection to Lahman baseball database in Aster 
conn = odbcDriverConnect(connection="driver={Aster ODBC Driver};
                         server=<dbhost>;port=2406;database=<dbname>;uid=<user>;pwd=<pw>")

showGraph(conn, "graph.films_vertices", "graph.films_edges")

}
}

