\docType{package}
\name{toaster}
\alias{toaster}
\alias{toaster-package}
\title{toaster: analytical and visualization toolbox for Teradata Aster Discovery
and Big Data Analytics Platform.}
\description{
toaster: analytical and visualization toolbox for Teradata
Aster Discovery and Big Data Analytics Platform.
}

