% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toaster.R
\docType{package}
\name{toaster}
\alias{toaster}
\alias{toaster-package}
\title{toaster: analytical and visualization toolbox with Teradata Aster big data
platform.}
\description{
toaster: analytical and visualization toolbox with Teradata Aster big data
platform.
}

